bump api and fix studio warnings

+++ app/src/main/java/ru/yourok/torrserve/server/local/ServerFile.kt
@@ -1,7 +1,6 @@
 package ru.yourok.torrserve.server.local
 
 import com.topjohnwu.superuser.Shell
-import ru.yourok.torrserve.BuildConfig
 import ru.yourok.torrserve.R
 import ru.yourok.torrserve.app.App
 import ru.yourok.torrserve.settings.Settings
@@ -15,7 +14,7 @@ class ServerFile : File(App.context.filesDir, "torrserver") {
         if (!exists())
             return
         synchronized(lock) {
-            Shell.Config.verboseLogging(BuildConfig.DEBUG)
+            //Shell.Config.verboseLogging(BuildConfig.DEBUG)
             val setspath = Settings.getTorrPath()
             val logfile = File(setspath, "torrserver.log").path
             if (shell == null) {
@@ -35,13 +34,11 @@ class ServerFile : File(App.context.filesDir, "torrserver") {
         if (!exists())
             return
         synchronized(lock) {
-            Shell.Config.verboseLogging(BuildConfig.DEBUG)
-            val result: Shell.Result = if (Shell.rootAccess())
+            //Shell.Config.verboseLogging(BuildConfig.DEBUG)
+            if (Shell.rootAccess())
                 Shell.su("killall -9 torrserver &").exec()
             else
                 Shell.sh("killall -9 torrserver &").exec()
-//            if (result.isSuccess)
-//                App.Toast(App.context.getString(R.string.server_stoped))
             shell = null
         }
     }
